Updates to the Square class and its use
The Square class now does the calculations for its on screen pixel coordinates itself in the render class, whereas before, it was rendering at the X and Y coordinates given to it in the main function.

The self.x and self.y properties of the Square class are now their coordinates in the checkerboard rather than their on screen coordinates.

This will be useful for piece kinging later on.