Added `selected` property to Piece class
The selected property is now what the render method will use to check what texture to use for the piece. Instead of setting the image for when the piece is selected, you can now set the selected property to true and the render method will do the rest. The move_piece method has been updated to comply with this new system. The game setup will now comply as well, defaulting to setting the selected property of the pieces to false.